DL_Projects/Landscape_Recognition/requirements.txt
kaggle==1.5.15
matplotlib==3.6.3
Pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
